Copter: move proximity log to DataFlash
